requirements.txt
undetected-chromedriver>=3.5.5
selenium>=4.9.0
requests>=2.31.0
pyyaml>=6.0
anthropic>=0.20.0
